New NorwegianNormalizationFilter. Refactored core functionality into utility class
