+++ lucene/analysis/common/src/java/org/apache/lucene/analysis/miscellaneous/ScandinavianNormalizationFilter.java
@@ -16,12 +16,13 @@
  */
 package org.apache.lucene.analysis.miscellaneous;
 
+import static org.apache.lucene.analysis.miscellaneous.ScandinavianNormalizer.ALL_FOLDINGS;
+
 import java.io.IOException;
 import java.util.Set;
 import org.apache.lucene.analysis.TokenFilter;
 import org.apache.lucene.analysis.TokenStream;
 import org.apache.lucene.analysis.tokenattributes.CharTermAttribute;
-import org.apache.lucene.analysis.util.StemmerUtil;
 
 /**
  * This filter normalize use of the interchangeable Scandinavian characters æÆäÄöÖøØ and folded
@@ -34,131 +35,36 @@ import org.apache.lucene.analysis.util.StemmerUtil;
  * <p>blåbærsyltetøj == blåbärsyltetöj == blaabaarsyltetoej but not blabarsyltetoj räksmörgås ==
  * ræksmørgås == ræksmörgaos == raeksmoergaas but not raksmorgas
  *
- * <p>You can choose which of the foldings to apply (aa, ao, ae, oe, oo) through a parameter.
+ * <p>There are also separate filters for Norwegian, Danish and Swedish with slightly differing
+ * settings
  *
  * @see ScandinavianFoldingFilter
  */
-public final class ScandinavianNormalizationFilter extends TokenFilter {
+public class ScandinavianNormalizationFilter extends TokenFilter {
 
-  /**
-   * Create the filter with default folding rules, backward compatible with all earlier versions
-   *
-   * @param input the TokenStream
-   */
+  private final ScandinavianNormalizer normalizer;
+
   public ScandinavianNormalizationFilter(TokenStream input) {
     super(input);
-    this.foldings = ALL_FOLDINGS;
+    this.normalizer = new ScandinavianNormalizer(ALL_FOLDINGS);
   }
 
-  /**
-   * Create the filter using custom folding rules.
-   *
-   * @param input the TokenStream
-   * @param foldings a Set of Foldings to apply (i.e. AE, OE, AA, AO, OO)
-   */
-  public ScandinavianNormalizationFilter(TokenStream input, Set<Foldings> foldings) {
+  // Constructor for use by sub classes
+  protected ScandinavianNormalizationFilter(
+      TokenStream input, Set<ScandinavianNormalizer.Foldings> foldings) {
     super(input);
-    this.foldings = foldings;
+    this.normalizer = new ScandinavianNormalizer(foldings);
   }
-
-  /** List of possible foldings that can be used when configuring the filter */
-  public enum Foldings {
-    AA,
-    AO,
-    AE,
-    OE,
-    OO
-  }
-
-  protected static final Set<Foldings> ALL_FOLDINGS =
-      Set.of(Foldings.AA, Foldings.AO, Foldings.OO, Foldings.AE, Foldings.OE);
 
   private final CharTermAttribute charTermAttribute = addAttribute(CharTermAttribute.class);
 
-  private static final char AA = '\u00C5'; // Å
-  private static final char aa = '\u00E5'; // å
-  private static final char AE = '\u00C6'; // Æ
-  private static final char ae = '\u00E6'; // æ
-  private static final char AE_se = '\u00C4'; // Ä
-  private static final char ae_se = '\u00E4'; // ä
-  private static final char OE = '\u00D8'; // Ø
-  private static final char oe = '\u00F8'; // ø
-  private static final char OE_se = '\u00D6'; // Ö
-  private static final char oe_se = '\u00F6'; // ö
-
-  private final Set<Foldings> foldings;
-
   @Override
-  public boolean incrementToken() throws IOException {
+  public final boolean incrementToken() throws IOException {
     if (!input.incrementToken()) {
       return false;
     }
-
-    char[] buffer = charTermAttribute.buffer();
-    int length = charTermAttribute.length();
-
-    int i;
-    for (i = 0; i < length; i++) {
-
-      if (buffer[i] == ae_se) {
-        buffer[i] = ae;
-
-      } else if (buffer[i] == AE_se) {
-        buffer[i] = AE;
-
-      } else if (buffer[i] == oe_se) {
-        buffer[i] = oe;
-
-      } else if (buffer[i] == OE_se) {
-        buffer[i] = OE;
-
-      } else if (length - 1 > i) {
-
-        if (buffer[i] == 'a'
-            && (foldings.contains(Foldings.AA) && (buffer[i + 1] == 'a' || buffer[i + 1] == 'A')
-                || foldings.contains(Foldings.AO)
-                    && (buffer[i + 1] == 'o' || buffer[i + 1] == 'O'))) {
-          length = StemmerUtil.delete(buffer, i + 1, length);
-          buffer[i] = aa;
-
-        } else if (buffer[i] == 'A'
-            && (foldings.contains(Foldings.AA) && (buffer[i + 1] == 'a' || buffer[i + 1] == 'A')
-                || foldings.contains(Foldings.AO)
-                    && (buffer[i + 1] == 'o' || buffer[i + 1] == 'O'))) {
-          length = StemmerUtil.delete(buffer, i + 1, length);
-          buffer[i] = AA;
-
-        } else if (buffer[i] == 'a'
-            && foldings.contains(Foldings.AE)
-            && (buffer[i + 1] == 'e' || buffer[i + 1] == 'E')) {
-          length = StemmerUtil.delete(buffer, i + 1, length);
-          buffer[i] = ae;
-
-        } else if (buffer[i] == 'A'
-            && foldings.contains(Foldings.AE)
-            && (buffer[i + 1] == 'e' || buffer[i + 1] == 'E')) {
-          length = StemmerUtil.delete(buffer, i + 1, length);
-          buffer[i] = AE;
-
-        } else if (buffer[i] == 'o'
-            && (foldings.contains(Foldings.OE) && (buffer[i + 1] == 'e' || buffer[i + 1] == 'E')
-                || foldings.contains(Foldings.OO)
-                    && (buffer[i + 1] == 'o' || buffer[i + 1] == 'O'))) {
-          length = StemmerUtil.delete(buffer, i + 1, length);
-          buffer[i] = oe;
-
-        } else if (buffer[i] == 'O'
-            && (foldings.contains(Foldings.OE) && (buffer[i + 1] == 'e' || buffer[i + 1] == 'E')
-                || foldings.contains(Foldings.OO)
-                    && (buffer[i + 1] == 'o' || buffer[i + 1] == 'O'))) {
-          length = StemmerUtil.delete(buffer, i + 1, length);
-          buffer[i] = OE;
-        }
-      }
-    }
-
-    charTermAttribute.setLength(length);
-
+    charTermAttribute.setLength(
+        normalizer.processToken(charTermAttribute.buffer(), charTermAttribute.length()));
     return true;
   }
 }
